use Byte to indicate Float in the SimilarityMatrix of recommender

+++ include/idmlib/resys/incremcf/IncrementalItemCF.h
@@ -64,7 +64,7 @@ public:
 
 private:
     ItemCoVisitation<CoVisitFreq> covisitation_;
-    SimilarityMatrix<uint32_t,float> similarity_;
+    SimilarityMatrix<uint32_t,uint8_t> similarity_;
     UserRecommendItem userRecommendItems_;
     size_t max_items_stored_for_each_user_;
 };
